.ipynb_checkpoints/requirements-checkpoint.txt
absl-py==2.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
contextlib2==21.6.0
contourpy==1.2.1
cycler==0.12.1
docker-pycreds==0.4.0
etils==1.5.2
filelock==3.15.4
flax==0.8.5
fonttools==4.53.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
jax==0.4.27
# jax-metal==0.1.0
jaxlib==0.4.27
Jinja2==3.1.4
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
ml-collections==0.1.1
ml-dtypes==0.4.0
mpmath==1.3.0
msgpack==1.0.8
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.24.3
opt-einsum==3.3.0
optax==0.2.3
orbax-checkpoint==0.5.20
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
protobuf==5.27.2
psutil==6.0.0
Pygments==2.18.0
pyparsing==3.1.2
PyQt5==5.15.10
# PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
pyvista==0.44.0
pyvistaqt==0.11.1
PyYAML==6.0.1
QtPy==2.4.1
requests==2.32.3
rich==13.7.1
scipy==1.13.1
scooby==0.10.0
sentry-sdk==2.9.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.13.0
tabulate==0.9.0
tensorstore==0.1.63
toolz==0.12.1
torch==2.3.1
typing_extensions==4.12.2
urllib3==2.2.2
vtk==9.3.1
wandb==0.17.4
zipp==3.19.2
